refactor(web): give each projects component its own file
components.tsx had grown to hold three unrelated things, and the only reason
they shared a module was the folder-icon hook between two of them. That hook
is now its own file and the rest follow lib/agents/components: one component
per file behind a barrel, so the import path callers use does not change when
a file is split again.

The three project components living elsewhere come home with them — the chat
list and context panel from sections/projects, the create modal from
sections/modals. sections/projects is now empty and gone.
