+++ web/src/lib/projects/components/ProjectContextPanel.tsx
