+++ web/src/lib/projects/components/index.ts
@@ -1,0 +1,9 @@
+export { useFolderIcon } from "@/lib/projects/components/useFolderIcon";
+export {
+  ProjectFolderButton,
+  type ProjectFolderButtonProps,
+} from "@/lib/projects/components/ProjectFolderButton";
+export { FoldedProjectsPopover } from "@/lib/projects/components/FoldedProjectsPopover";
+export { default as ProjectChatSessionList } from "@/lib/projects/components/ProjectChatSessionList";
+export { default as ProjectContextPanel } from "@/lib/projects/components/ProjectContextPanel";
+export { default as CreateProjectModal } from "@/lib/projects/components/CreateProjectModal";
