+++ web/src/sections/sidebar/AppSidebar.tsx
@@ -43,7 +43,7 @@ import {
   FoldedProjectsPopover,
   ProjectFolderButton,
 } from "@/lib/projects/components";
-import CreateProjectModal from "@/sections/modals/CreateProjectModal";
+import { CreateProjectModal } from "@/lib/projects/components";
 import { MoveCustomAgentChatModal } from "@/lib/agents/components";
 import { useProjectsContext } from "@/lib/projects/providers";
 import { removeChatSessionFromProject } from "@/lib/projects/svc";
