+++ web/src/views/AppPage.tsx
@@ -53,11 +53,11 @@ import FederatedOAuthModal from "@/components/chat/FederatedOAuthModal";
 import ChatScrollContainer, {
   ChatScrollContainerHandle,
 } from "@/sections/chat/ChatScrollContainer";
-import ProjectContextPanel from "@/sections/projects/ProjectContextPanel";
+import { ProjectContextPanel } from "@/lib/projects/components";
 import { useProjectsContext } from "@/lib/projects/providers";
 import { useActiveProject, useProjects } from "@/lib/projects/hooks";
 import { getProjectTokenCount } from "@/lib/projects/svc";
-import ProjectChatSessionList from "@/sections/projects/ProjectChatSessionList";
+import { ProjectChatSessionList } from "@/lib/projects/components";
 import { cn } from "@opal/utils";
 import Suggestions from "@/sections/Suggestions";
 import OnboardingFlow from "@/sections/onboarding/OnboardingFlow";
